user_idに基づいてビッグファイブデータを取り出して会話のプロンプトに使用できるようにしました。
